Set up code for fetching temperature data
For this project, I decided to use d3.json to fetch the data from the
web instead of downloading it and importing via ES6 modules. Just for a
change.

Also, it looks like eslint allows a line to contain more than 80 chars
if it contains a giant string (like a long URL).