Plot the data points along year and month axes
This is just plotting the data points along the year and month axes.
I've yet to visualize the temperatures.

If you look closely, the rectangles for 30-day months overlap with the
rectangles below them. And there's a glaring overlap between February
and March. It's because the vertical scale is based on a continuous date
scale, so it represents the length of months appropriately. However
what I'd like to do is to treat the vertical scale as a discrete scale
with 12 ticks. But it's not a priority.