Display tooltip on cell hover
This time I used the d3-tip library instead of writing my own tooltip
code. Arguably much neater, though the HTML template is still a mess.

I want to put a little arrow on the bottom of the tooltip to point on
its respective cell, but I couldn't get it right.