Add legend labels
To accommodate the labels at the bottom, I increased the bottom padding
and decreased the top padding.

Producing suitable labels turned out to be simple. I just had to get the
span/extent of the temperature range and divide by 11 (because the color
scheme has 11 colors in it), then pass that result to d3.range (along
with the temperature extremes) and it produced a nice array of
temperatures that match the colors.