Change month scale from to band scale
This gives me evenly-sized ticks on the month axis. The problem with
scaleTime was that the tick sizes were not even (due to the face that
months have different number of days), and so that causes overlap
between cells (whose heights are based on a crude calculation).

I also removed the second `call` call on the month axis, because I no
longer have to shift the labels in the code.

This change requires me to code in the month names (because scaleBand
requires a discrete domain).

Thanks to @honmanyau for info on `.tickSize()`!